Fix for CLJ-886, along with test illustrating the problem
Original reported problem was with do-copy method on arg types
InputStream and Writer.  Changes to function data-fixture and test
test-copy-encodings in clojure.java.io's test suite now cause the
problem to occur, in the absence of this fix to do-copy.

While fixing that one and adding tests, it was also discovered that
do-copy with arg types Reader and OutputStream had a bug if the
encoding was UTF-16, because Java's .getBytes returns a byte-order
mark (Unicode code point U+FEFF) at the beginning of the byte array
every time it is called.  Thus if the file being copied was larger
than one buffer's worth of data, a new byte-order mark was written to
the copy in between each buffer's worth.  With the correction, only a
single byte-order mark is written at the beginning.

Also enhanced the tests test-spit-and-slurp and test-input-stream to
use non-ASCII characters.  No bugs were found as a result of these
test enhancements.

Signed-off-by: Stuart Halloway <stu@thinkrelevance.com>